only clear embedding folder after validation

+++ src/train_hooks.py
@@ -380,8 +380,9 @@ class BatchDumpHook(tf.train.SessionRunHook):
         if not os.path.exists(self.out_dir):
             os.makedirs(self.out_dir)
         else:
-            shutil.rmtree(self.out_dir)
-            os.makedirs(self.out_dir)
+            if not train:
+                shutil.rmtree(self.out_dir)
+                os.makedirs(self.out_dir)
 
     def get_feature_folder_path(self):
         return self.out_dir
@@ -569,9 +570,6 @@ class PredictionHook(tf.train.SessionRunHook):
         )
         if not os.path.exists(self.out_dir):
             os.makedirs(self.out_dir)
-        else:
-            shutil.rmtree(self.out_dir)
-            os.makedirs(self.out_dir)
 
         self.streamer = streamer
         self.target_label = target_label
@@ -749,9 +747,6 @@ class PredictionRobustnessHook(tf.train.SessionRunHook):
             'prediction_robustness_' + str(epoch)
         )
         if not os.path.exists(self.out_dir):
-            os.makedirs(self.out_dir)
-        else:
-            shutil.rmtree(self.out_dir)
             os.makedirs(self.out_dir)
 
         self.streamer = streamer
